requirements.txt
numpy==1.19.0
python-binance==0.7.5
matplotlib==3.2.1
TA-Lib==0.4.18